Added option for installing software.

+++ setup.sh
@@ -16,6 +16,8 @@ RCNFS_DATASETS_EXPORT_DIR=$2
 RCXX_BACKUP_DIR=$3
 # Account in which various software should be setup.
 USERNAME=$4
+# Whether or not to install additional software beyond the basics.
+INSTALL_SOFTWARE=$5
 
 # === Paarameters decided by this script. ===
 # Directory where the NFS partition will be mounted on NFS clients
@@ -68,40 +70,44 @@ apt-get update
 apt-get --assume-yes install mosh vim tmux pdsh tree axel htop ctags
 # NFS
 apt-get --assume-yes install nfs-kernel-server nfs-common
-# Java
-apt-get install --assume-yes software-properties-common
-add-apt-repository --yes ppa:webupd8team/java
-apt-get update
-echo oracle-java8-installer shared/accepted-oracle-license-v1-1 select true | sudo /usr/bin/debconf-set-selections
-apt-get install --assume-yes oracle-java8-installer
-# Maven
-wget http://www-us.apache.org/dist/maven/maven-3/3.5.2/binaries/apache-maven-3.5.2-bin.tar.gz
-tar -xvzf apache-maven-3.5.2-bin.tar.gz
-mv apache-maven-3.5.2 /usr/lib
 
-cat >> /etc/profile.d/maven.sh <<EOM
+if [ "$INSTALL_SOFTWARE" == "True" ]
+then
+  # Java
+  apt-get install --assume-yes software-properties-common
+  add-apt-repository --yes ppa:webupd8team/java
+  apt-get update
+  echo oracle-java8-installer shared/accepted-oracle-license-v1-1 select true | sudo /usr/bin/debconf-set-selections
+  apt-get install --assume-yes oracle-java8-installer
+  # Maven
+  wget http://www-us.apache.org/dist/maven/maven-3/3.5.2/binaries/apache-maven-3.5.2-bin.tar.gz
+  tar -xvzf apache-maven-3.5.2-bin.tar.gz
+  mv apache-maven-3.5.2 /usr/lib
+
+  cat >> /etc/profile.d/maven.sh <<EOM
 export PATH=/usr/lib/apache-maven-3.5.2/bin:$PATH
 EOM
-chmod ugo+x /etc/profile.d/maven.sh
-# cpupower, hugepages, msr-tools (for rdmsr), i7z
-apt-get --assume-yes install linux-tools-common linux-tools-${KERNEL_RELEASE} \
+  chmod ugo+x /etc/profile.d/maven.sh
+  # cpupower, hugepages, msr-tools (for rdmsr), i7z
+  apt-get --assume-yes install linux-tools-common linux-tools-${KERNEL_RELEASE} \
         hugepages cpuset msr-tools i7z
-# Dependencies to build the Linux perf tool
-apt-get --assume-yes install systemtap-sdt-dev libunwind-dev libaudit-dev \
+  # Dependencies to build the Linux perf tool
+  apt-get --assume-yes install systemtap-sdt-dev libunwind-dev libaudit-dev \
         libgtk2.0-dev libperl-dev binutils-dev liblzma-dev libiberty-dev
-# Install RAMCloud dependencies
-apt-get --assume-yes install build-essential git-core doxygen libpcre3-dev \
+  # Install RAMCloud dependencies
+  apt-get --assume-yes install build-essential git-core doxygen libpcre3-dev \
         protobuf-compiler libprotobuf-dev libcrypto++-dev libevent-dev \
         libboost-all-dev libgtest-dev libzookeeper-mt-dev zookeeper \
         libssl-dev default-jdk ccache
-# Mellanox OFED (Note: Reboot required after installing this).
-apt-get --assume-yes install tk8.4 chrpath graphviz tcl8.4 libgfortran3 dkms \
+  # Mellanox OFED (Note: Reboot required after installing this).
+  apt-get --assume-yes install tk8.4 chrpath graphviz tcl8.4 libgfortran3 dkms \
         tcl pkg-config gfortran curl libnl1 quilt dpatch swig tk python-libxml2
-OS_VER="ubuntu`lsb_release -r | cut -d":" -f2 | xargs`"
-MLNX_OFED="MLNX_OFED_LINUX-3.4-1.0.0.0-$OS_VER-x86_64"
-axel -n 8 -q http://www.mellanox.com/downloads/ofed/MLNX_OFED-3.4-1.0.0.0/$MLNX_OFED.tgz
-tar xzf $MLNX_OFED.tgz
-./$MLNX_OFED/mlnxofedinstall --force --without-fw-update
+  OS_VER="ubuntu`lsb_release -r | cut -d":" -f2 | xargs`"
+  MLNX_OFED="MLNX_OFED_LINUX-3.4-1.0.0.0-$OS_VER-x86_64"
+  axel -n 8 -q http://www.mellanox.com/downloads/ofed/MLNX_OFED-3.4-1.0.0.0/$MLNX_OFED.tgz
+  tar xzf $MLNX_OFED.tgz
+  ./$MLNX_OFED/mlnxofedinstall --force --without-fw-update
+fi
 
 # === Configuration settings for all machines ===
 # Make vim the default editor.
@@ -207,8 +213,11 @@ then
 fi
 EOM
 
-  # Execute all user-specific setup in user's shared folder using rcmaster.
-  sudo --login -u $USERNAME $SCRIPTPATH/user-setup.sh $RCXX_BACKUP_DIR
+  if [ "$INSTALL_SOFTWARE" == "True" ]
+  then
+    # Execute all user-specific setup in user's shared folder using rcmaster.
+    sudo --login -u $USERNAME $SCRIPTPATH/user-setup.sh $RCXX_BACKUP_DIR
+  fi
 fi
 
 # RCXX machines specific configuration.
@@ -230,43 +239,46 @@ then
 * hard memlock unlimited
 EOM
 
-  # Enable cpuset functionality on rc machines. This is also optional, RAMCloud
-  # will work without this. TODO: Check whether or not this is actually
-  # necessary after installing the cpuset package on these machines.
-  if [ ! -d "/sys/fs/cgroup/cpuset" ]; then
-    mount -t tmpfs cgroup_root /sys/fs/cgroup
-    mkdir /sys/fs/cgroup/cpuset
-    mount -t cgroup cpuset -o cpuset /sys/fs/cgroup/cpuset/
+  if [ "$INSTALL_SOFTWARE" == "True" ]
+  then
+    # Enable cpuset functionality on rc machines. This is also optional, RAMCloud
+    # will work without this. TODO: Check whether or not this is actually
+    # necessary after installing the cpuset package on these machines.
+    if [ ! -d "/sys/fs/cgroup/cpuset" ]; then
+      mount -t tmpfs cgroup_root /sys/fs/cgroup
+      mkdir /sys/fs/cgroup/cpuset
+      mount -t cgroup cpuset -o cpuset /sys/fs/cgroup/cpuset/
+    fi
+
+    # Enable hugepage support for DPDK: 
+    # http://dpdk.org/doc/guides/linux_gsg/sys_reqs.html
+    # The changes will take effects after reboot. m510 is not a NUMA machine.
+    # Reserve 1GB hugepages via kernel boot parameters
+    kernel_boot_params="default_hugepagesz=1G hugepagesz=1G hugepages=8"
+
+    # Disable intel_idle driver to gain control over C-states (this driver will
+    # most ignore any other BIOS setting and kernel parameters). Then limit
+    # available C-states to C1 by "idle=halt". Or more aggressively, keep
+    # processors in C0 even when they are idle by "idle=poll".
+    kernel_boot_params+=" intel_idle.max_cstate=0 idle=poll"
+
+    # Isolate certain cpus from kernel scheduling and put them into full
+    # dynticks mode (need reboot to take effect)
+    #isolcpus="2"
+    #kernel_boot_params+=" isolcpus=$isolcpus nohz_full=$isolcpus rcu_nocbs=$isolcpus"
+
+    # Enable perf taken branch stack sampling (i.e. "perf record -b ...")
+    #kernel_boot_params+=" lapic"
+
+    # Update GRUB with our kernel boot parameters
+    sed -i "s/GRUB_CMDLINE_LINUX_DEFAULT=\"/GRUB_CMDLINE_LINUX_DEFAULT=\"$kernel_boot_params /" /etc/default/grub
+    update-grub
+    # TODO: Verify that these options actually work as expected.
+    # http://www.breakage.org/2013/11/15/nohz_fullgodmode/
+
+    # Note: We will reboot the rc machines at the end of this script so that the
+    # kernel parameter changes can take effect.
   fi
-
-  # Enable hugepage support for DPDK: 
-  # http://dpdk.org/doc/guides/linux_gsg/sys_reqs.html
-  # The changes will take effects after reboot. m510 is not a NUMA machine.
-  # Reserve 1GB hugepages via kernel boot parameters
-  kernel_boot_params="default_hugepagesz=1G hugepagesz=1G hugepages=8"
-
-  # Disable intel_idle driver to gain control over C-states (this driver will
-  # most ignore any other BIOS setting and kernel parameters). Then limit
-  # available C-states to C1 by "idle=halt". Or more aggressively, keep
-  # processors in C0 even when they are idle by "idle=poll".
-  kernel_boot_params+=" intel_idle.max_cstate=0 idle=poll"
-
-  # Isolate certain cpus from kernel scheduling and put them into full
-  # dynticks mode (need reboot to take effect)
-  #isolcpus="2"
-  #kernel_boot_params+=" isolcpus=$isolcpus nohz_full=$isolcpus rcu_nocbs=$isolcpus"
-
-  # Enable perf taken branch stack sampling (i.e. "perf record -b ...")
-  #kernel_boot_params+=" lapic"
-
-  # Update GRUB with our kernel boot parameters
-  sed -i "s/GRUB_CMDLINE_LINUX_DEFAULT=\"/GRUB_CMDLINE_LINUX_DEFAULT=\"$kernel_boot_params /" /etc/default/grub
-  update-grub
-  # TODO: Verify that these options actually work as expected.
-  # http://www.breakage.org/2013/11/15/nohz_fullgodmode/
-
-  # Note: We will reboot the rc machines at the end of this script so that the
-  # kernel parameter changes can take effect.
 fi
 
 # Mark that setup has finished. This script is actually run again after a
@@ -277,5 +289,8 @@ fi
 # Reboot required on rc machines for kernel parameter changes to take effect.
 if [[ $(hostname --short) =~ ^rc[0-9][0-9]$ ]]
 then
-  reboot
+  if [ "$INSTALL_SOFTWARE" == "True" ]
+  then
+    reboot
+  fi
 fi
